+++ xonsh/tools.py
@@ -38,6 +38,7 @@ import traceback
 import typing as tp
 import warnings
 
+
 # adding imports from further xonsh modules is discouraged to avoid circular
 # dependencies
 from xonsh import __version__
@@ -51,6 +52,16 @@ from xonsh.platform import (
     os_environ,
     pygments_version_info,
 )
+
+from contextlib import contextmanager
+
+
+@contextmanager
+def chdir(adir):
+    old_dir = os.getcwd()
+    os.chdir(adir)
+    yield
+    os.chdir(old_dir)
 
 
 @functools.lru_cache(1)
